Add paginate and regenerate acts

+++ db/migrate/20161226191952_create_acts.rb
@@ -2,7 +2,7 @@ class CreateActs < ActiveRecord::Migration
   def change
     create_table :acts do |t|
       t.string :name
-      t.integer :AMO
+      t.float :AMO
 
       t.timestamps null: false
     end
